requirements.txt
Django==5.1.4
psycopg===3.2.3
psycopg-binary==3.2.3
psycopg-pool===3.2.4
channels==4.2.0
daphne==4.1.2
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
priority==2.0.0
pydantic==2.10.4
pydantic-settings==2.7.1
django-cors-headers==4.6.0
celery==5.4.0
channels-redis==4.2.1
django-redis==5.4.0
shapely==2.0.6
confluent-kafka==2.8.0
ujson==5.10.0
pydantic-extra-types==2.10.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
colorlog==6.9.0
asgiref==3.8.1
